% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbd2015}
\alias{gbd2015}
\title{Data from the Global Burden of Disease Study 2015}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{cause_name}{medical cause / disease}
  \item{cause_id}{individual id for each disease}
  \item{metric_name}{class: number, rate or percent}
  \item{acute_incidence}{acute incidence rate}
  \item{chronic_prevalence}{chronic prevalence rate}
  \item{DALYs}{disability adjusted life year - describes the numer of years lost due to ill-health, disability or early death}
  \item{deaths}{number of deaths in the population}
  \item{incidence}{number of incident cases in the population }
  \item{prevalence}{prevalence rate}
  \item{YLDs}{years lost due to disability}
  \item{YLLs}{years of life lost due to dying early}
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
gbd2015
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
