% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generate a WizRd Report}
\usage{
generate(df)
}
\arguments{
\item{df}{Dataframe to generate the Report for}
}
\description{
Generate a WizRd Report
}
\examples{
generate(cars)
}
