% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_child.R
\name{render_child}
\alias{render_child}
\title{Render a child R Markdown document inside a parent R Markdown document.
Has to run inside a .Rmd document!}
\usage{
render_child(path)
}
\arguments{
\item{path}{Path to the .Rmd file (relative to current document)}
}
\description{
Render a child R Markdown document inside a parent R Markdown document.
Has to run inside a .Rmd document!
}
\examples{
```{r, results='asis'}
render.child("_general.Rmd")
```
}
