% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generate a WizRd Report}
\usage{
generate(df, pattern = NULL, template = NULL, output_format = "all")
}
\arguments{
\item{df}{Dataframe to generate the Report for}

\item{pattern}{Pattern / Regex to use for subsettings the dataframe}

\item{template}{Path to .Rmd template}

\item{output_format}{Format to generate for the report}
}
\description{
Generate a WizRd Report
}
\examples{
Generate a report for the cars dataset
generate(cars)

# Generate a report for all columns containing "DALY" or "YLD"
generate(gbd2015, pattern = "DALY|YLD")

# Generate only a html report
generate(gbd2015, output_format = "html_document")
}
